+++ src/frontend/packages/core/src/app.component.ts
@@ -1,13 +1,13 @@
-import { AfterContentInit, Component, HostBinding, OnDestroy, OnInit, Inject } from '@angular/core';
+import { DOCUMENT } from '@angular/common';
+import { AfterContentInit, Component, HostBinding, Inject, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { create } from 'rxjs-spy';
 
 import { AuthOnlyAppState } from '../../store/src/app-state';
-import { ThemeService } from './core/theme.service';
+import { ThemeService } from '../../store/src/theme.service';
 import { environment } from './environments/environment';
 import { LoggedInService } from './logged-in.service';
-import { DOCUMENT } from '@angular/common';
 
 @Component({
   selector: 'app-root',
