+++ src/frontend/packages/core/src/features/user-profile/profile-info/profile-info.component.ts
@@ -6,8 +6,8 @@ import { map } from 'rxjs/operators';
 import { SetPollingEnabledAction, SetSessionTimeoutAction } from '../../../../../store/src/actions/dashboard-actions';
 import { DashboardOnlyAppState } from '../../../../../store/src/app-state';
 import { selectDashboardState } from '../../../../../store/src/selectors/dashboard.selectors';
+import { ThemeService } from '../../../../../store/src/theme.service';
 import { UserProfileInfo } from '../../../../../store/src/types/user-profile.types';
-import { ThemeService } from '../../../core/theme.service';
 import { UserProfileService } from '../../../core/user-profile.service';
 import { UserService } from '../../../core/user.service';
 import { SetGravatarEnabledAction } from './../../../../../store/src/actions/dashboard-actions';
